// types.ts
export interface PurchaseOrderProduct {
  inventoryId: string;
  productName: string;
  buyingPrice: number;
  purchaseQuantity: number;
  productCode?: string;
  productStatus?: "pending" | "received" | "partially_received" | string;
  _id?: string;
}

export interface ApiPurchaseOrder {
  _id: string;
  supplierId: any;
  products: PurchaseOrderProduct[];
  status: string;
  note: string;
  totalAmount: number;
  paymentType?: "paid" | "credit";
  paidAmount?: number;
  paymentStatus?: "paid" | "partial" | "unpaid" | string;
  remainingBalance?: number;
  dueDate?: string | null;
  createdAt: string;
  updatedAt: string;
  __v?: number;
  id?: string;
  poNumber: string;
  totalRemainingQuantity?: number;
  purchasedBy?: any;
  isDeleted?: boolean;
  deletedAt?: string;
}
// ...
export interface Supplier {
  id: string;
  _id?: string;
  supplierName: string;
  contactNumber: string;
  township?: string;
  address?: string;
  isDeleted?: boolean;
  deletedAt?: string | null;
  createdAt?: string;
  updatedAt?: string;
}